// apps/pwabuilder-chromium-extension/src/interfaces/manifest.ts
export interface Manifest {
  background_color?: string;
  categories?: string[];
  description?: string;
  dir?: "auto" | "ltr" | "rtl" | string;
  display?: "fullscreen" | "standalone" | "minimal-ui" | "browser";
  display_override?: "fullscreen" | "standalone" | "minimal-ui" | "browser" | "window-controls-overlay";
  iarc_rating_id?: string;
  iconBlobUrls?: string[];
  icons?: Icon[];
  lang?: string | undefined;
  name?: string | undefined;
  orientation?:
    | "any"
    | "natural"
    | "landscape"
    | "portrait"
    | "portrait-primary"
    | "portrait-secondary"
    | "landscape-primary"
    | "landscape-secondary";
  prefer_related_applications?: boolean;
  protocol_handlers?: ProtocolHandler[];
  related_applications?: RelatedApplication[];
  scope?: string;
  screenshots?: Screenshot[];
  share_target?: ShareTarget;
  short_name?: string;
  shortcuts?: ShortcutItem[];
  start_url?: string;
  theme_color?: string;
  generated?: boolean;

  // for custom properties as well as using object notations: manifest[key]
  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
  // @ts-ignore - accomodate custom entries... these can be a pain
  [key: string]: string | boolean | undefined | Array<any> | any;
}

export interface ProtocolHandler {
  protocol: string;
  url: string;
}

export interface ShortcutItem {
  name: string;
  url: string;
  description?: string;
  short_name?: string;
  icons?: Icon[];
}

export interface Icon {
  src: string;
  generated?: boolean;
  type?: string;
  sizes?: string;
  purpose?: "any" | "maskable" | "monochrome";
  label?: string;
}

export interface Screenshot extends Icon {
  platform?:
    | "narrow"
    | "wide"
    | "android"
    | "chromeos"
    | "ios"
    | "kaios"
    | "macos"
    | "windows"
    | "xbox"
    | "chrome_web_store"
    | "play"
    | "itunes"
    | "microsoft-inbox"
    | "microsoft-store";
}

export interface RelatedApplication {
  platform: string;
  url?: string | null;
  id?: string | null;
  min_version?: string | null;
  fingerprints?: Fingerprint[];
}

export interface Fingerprint {
  type: string;
  value: string;
}

export interface ShareTarget {
  action?: string;
  method?: string;
  enctype?: string;
  params?: ShareTargetParams;
}
// ...
export interface ManifestContext {
  siteUrl: string;
  manifestUrl: string;
  manifest: Manifest;
  initialManifest: Manifest;
  isGenerated: boolean;
  isEdited: boolean;
}

export interface RawTestResult {
  manifest: Array<TestResult> | boolean;
  service_worker: Array<TestResult>;
  security: Array<TestResult>;
}

export interface TestResult {
  infoString: string;
  result: boolean;
  category: string;
  description?: string;
}